+++ cognite/experimental/_api/geospatial.py
@@ -1,23 +1,28 @@
 import functools
+import numbers
 from typing import Any, Dict, List, Union
 
 from cognite.client._api_client import APIClient
 
-from cognite.experimental.data_classes.geospatial import Feature, FeatureList, FeatureType, FeatureTypeList
+from cognite.experimental.data_classes.geospatial import (
+    CoordinateReferenceSystem,
+    CoordinateReferenceSystemList,
+    Feature,
+    FeatureList,
+    FeatureType,
+    FeatureTypeList,
+)
 
 
 class ExperimentalGeospatialAPI(APIClient):
 
     X_COGNITE_DOMAIN = "x-cognite-domain"
 
-    _RESOURCE_PATH = "/spatial/featuretypes"
-    _LIST_CLASS = FeatureTypeList
-    _ASSERT_CLASSES = False
-
     _cognite_domain = None
 
-    def _feature_resource_path(self, feature_type: FeatureType):
-        return self._RESOURCE_PATH + "/" + feature_type.external_id + "/features"
+    @staticmethod
+    def _feature_resource_path(feature_type: FeatureType):
+        return f"/spatial/featuretypes/{feature_type.external_id}/features"
 
     def _with_cognite_domain(func):
         @functools.wraps(func)
@@ -40,8 +45,8 @@ class ExperimentalGeospatialAPI(APIClient):
     @_with_cognite_domain
     def create_feature_types(
         self, feature_type: Union[FeatureType, List[FeatureType]]
-    ) -> Union[FeatureType, List[FeatureType]]:
-        """`Creates feature types`_
+    ) -> Union[FeatureType, FeatureTypeList]:
+        """`Creates feature types`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/createFeatureTypes>
 
         Args:
@@ -63,11 +68,11 @@ class ExperimentalGeospatialAPI(APIClient):
                 ... ]
                 >>> res = c.geospatial.create_feature_types(feature_types)
         """
-        return self._create_multiple(items=feature_type)
+        return self._create_multiple(items=feature_type, cls=FeatureTypeList, resource_path="/spatial/featuretypes")
 
     @_with_cognite_domain
     def delete_feature_types(self, external_id: Union[str, List[str]] = None) -> None:
-        """`Delete one or more feature type`_
+        """`Delete one or more feature type`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/deleteFeatureTypes>
 
         Args:
@@ -84,11 +89,11 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> c.geospatial.delete_feature_types(external_id=["wells", "pipelines"])
         """
-        return self._delete_multiple(external_ids=external_id, wrap_ids=True)
+        return self._delete_multiple(external_ids=external_id, wrap_ids=True, resource_path="/spatial/featuretypes")
 
     @_with_cognite_domain
     def list_feature_types(self) -> FeatureTypeList:
-        """`List feature types`_
+        """`List feature types`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/listFeatureTypes>
 
         Returns:
@@ -103,11 +108,11 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> for feature_type in c.geospatial.list_feature_types():
                 ...     feature_type # do something with the feature type definition
         """
-        return self._list(method="POST")
+        return self._list(method="POST", cls=FeatureTypeList, resource_path="/spatial/featuretypes")
 
     @_with_cognite_domain
     def retrieve_feature_types(self, external_id: Union[str, List[str]] = None) -> FeatureTypeList:
-        """`Retrieve feature types`_
+        """`Retrieve feature types`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/getFeatureTypesByIds>
 
         Args:
@@ -124,13 +129,15 @@ class ExperimentalGeospatialAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> res = c.geospatial.retrieve_feature_types(external_id="1")
         """
-        return self._retrieve_multiple(wrap_ids=True, external_ids=external_id)
+        return self._retrieve_multiple(
+            wrap_ids=True, external_ids=external_id, cls=FeatureTypeList, resource_path="/spatial/featuretypes"
+        )
 
     @_with_cognite_domain
     def create_features(
         self, feature_type: FeatureType, feature: Union[Feature, List[Feature]]
-    ) -> Union[Feature, List[Feature]]:
-        """`Creates features`_
+    ) -> Union[Feature, FeatureList]:
+        """`Creates features`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/createFeatures>
 
         Args:
@@ -155,7 +162,7 @@ class ExperimentalGeospatialAPI(APIClient):
 
     @_with_cognite_domain
     def delete_features(self, feature_type: FeatureType, external_id: Union[str, List[str]] = None) -> None:
-        """`Delete one or more feature`_
+        """`Delete one or more feature`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/deleteFeatures>
 
         Args:
@@ -179,7 +186,7 @@ class ExperimentalGeospatialAPI(APIClient):
 
     @_with_cognite_domain
     def retrieve_features(self, feature_type: FeatureType, external_id: Union[str, List[str]] = None) -> FeatureList:
-        """`Retrieve features`_
+        """`Retrieve features`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/getFeaturesByIds>
 
         Args:
@@ -205,7 +212,7 @@ class ExperimentalGeospatialAPI(APIClient):
 
     @_with_cognite_domain
     def update_features(self, feature_type: FeatureType, feature: Union[Feature, List[Feature]]) -> FeatureList:
-        """`Update features`_
+        """`Update features`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/updateFeatures>
 
         Args:
@@ -233,7 +240,7 @@ class ExperimentalGeospatialAPI(APIClient):
 
     @_with_cognite_domain
     def search_features(self, feature_type: FeatureType, filter: Dict[str, Any], limit: int = 100) -> FeatureList:
-        """`Search for features`_
+        """`Search for features`
         <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/searchFeatures>
 
         Args:
@@ -258,6 +265,98 @@ class ExperimentalGeospatialAPI(APIClient):
         """
         resource_path = self._feature_resource_path(feature_type) + "/search"
         cls = FeatureList
-        resource_path = resource_path or self._RESOURCE_PATH
+        resource_path = resource_path
         res = self._post(url_path=resource_path, json={"filter": filter, "limit": limit},)
         return cls._load(res.json()["items"], cognite_client=self._cognite_client)
+
+    def get_coordinate_reference_systems(self, srids: Union[int, List[int]] = None) -> CoordinateReferenceSystemList:
+        """`Get Coordinate Reference Systems`
+        <https://pr-1323.specs.preview.cogniteapp.com/v1.json.html#operation/getCRS>
+
+        Args:
+            srids: (Union[int, List[int]]): SRID or list of SRIDs
+
+        Returns:
+            CoordinateReferenceSystemList: Requested CRSs.
+
+        Examples:
+
+            Get two CRS definitions:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> crs = c.geospatial.get_coordinate_reference_systems(srids=[4326, 4327])
+        """
+        if isinstance(srids, numbers.Integral):
+            srids = [srids]
+
+        res = self._post(url_path="/spatial/crs/byids", json={"items": [{"srid": srid} for srid in srids]})
+        return CoordinateReferenceSystemList._load(res.json()["items"], cognite_client=self._cognite_client)
+
+    def list_coordinate_reference_systems(self, only_custom=False) -> CoordinateReferenceSystemList:
+        """`Retrieve features`
+
+        Args:
+            only_custom: list only custom CRSs or not
+
+        Returns:
+            CoordinateReferenceSystemList: list of CRSs.
+
+        Examples:
+
+            Fetch all custom CRSs:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> crs = c.geospatial.list_coordinate_reference_systems(only_custom=True)
+        """
+        res = self._get(url_path="/spatial/crs/list", params={"filterCustom": only_custom})
+        return CoordinateReferenceSystemList._load(res.json()["items"], cognite_client=self._cognite_client)
+
+    def create_coordinate_reference_systems(
+        self, crs: Union[CoordinateReferenceSystem, List[CoordinateReferenceSystem]]
+    ) -> CoordinateReferenceSystemList:
+        """`Create Coordinate Reference System`
+
+        Args:
+            crs: a CoordinateReferenceSystem or a list of CoordinateReferenceSystem
+
+        Returns:
+            CoordinateReferenceSystemList: list of CRSs.
+
+        Examples:
+
+            Create a custom CRS:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> custom_crs = CoordinateReferenceSystem(srid = 121111, wkt="wkt", proj_string="proj")
+                >>> crs = c.geospatial.create_coordinate_reference_systems(custom_crs)
+        """
+        if isinstance(crs, CoordinateReferenceSystem):
+            crs = [crs]
+
+        res = self._post(url_path="/spatial/crs", json={"items": [it.dump(camel_case=True) for it in crs]})
+        return CoordinateReferenceSystemList._load(res.json()["items"], cognite_client=self._cognite_client)
+
+    def delete_coordinate_reference_systems(self, srids: Union[int, List[int]] = None) -> None:
+        """`Delete Coordinate Reference System`
+
+        Args:
+            srids: (Union[int, List[int]]): SRID or list of SRIDs
+
+        Returns:
+            None
+
+        Examples:
+
+            Delete a custom CRS:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> crs = c.geospatial.delete_coordinate_reference_systems(srids=[121111])
+        """
+        if isinstance(srids, numbers.Integral):
+            srids = [srids]
+
+        self._post(url_path="/spatial/crs/delete", json={"items": [{"srid": srid} for srid in srids]})