+++ cognite/experimental/_api/transformation_jobs.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import List, Optional
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
@@ -89,3 +89,24 @@ class TransformationJobsAPI(APIClient):
         return self._list(
             method="GET", limit=LIST_LIMIT_CEILING, resource_path=url_path, cls=TransformationJobMetricList
         )
+
+    def retrieve_multiple(self, ids: List[int], ignore_unknown_ids: bool = False) -> TransformationJobList:
+        """`Retrieve multiple transformation jobs by id. <https://docs.cognite.com/api/playground/#operation/getTransformationJob>`_
+
+        Args:
+            ids (List[int]): Job internal Ids
+            ignore_unknown_ids (bool): Ignore IDs that are not found rather than throw an exception.
+
+        Returns:
+            TransformationJobList: Requested transformation jobs.
+
+        Examples:
+
+            Get jobs by id::
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> res = c.transformations.jobs.retrieve_multiple(ids=[1, 2, 3])
+        """
+        utils._auxiliary.assert_type(ids, "id", [List], allow_none=True)
+        return self._retrieve_multiple(ids=ids, ignore_unknown_ids=ignore_unknown_ids, wrap_ids=True)